FM-8449 - convert vsphere to litmus
add github actions
