+++ .fixtures.yml
@@ -1,4 +1,8 @@
 fixtures:
   symlinks:
     vsphere: "#{source_dir}"
+  repositories:
+    facts: 'git://github.com/puppetlabs/puppetlabs-facts.git'
+    puppet_agent: 'git://github.com/puppetlabs/puppetlabs-puppet_agent.git'
+    provision: 'git://github.com/puppetlabs/provision.git'
