+++ spec/spec_helper_acceptance.rb
@@ -1,192 +1,82 @@
-require 'mustache'
-require 'open3'
-require 'rbvmomi'
-require 'retries'
+# frozen_string_literal: true
 
-# This exception is used to signal expected continuations when waiting for events on the vCenter
-class NotFinished < RuntimeError
-end
+require 'serverspec'
+require 'puppet_litmus'
+require 'spec_helper_acceptance_local' if File.file?(File.join(File.dirname(__FILE__), 'spec_helper_acceptance_local.rb'))
+include PuppetLitmus
 
-class PuppetManifest < Mustache
-  def initialize(file, config)
-    @template_file = File.join(Dir.getwd, 'spec', 'acceptance', 'fixtures', file)
-    config.each do |key, value|
-      config_value = self.class.to_generalized_data(value)
-      instance_variable_set("@#{key}".to_sym, config_value)
-      self.class.send(:attr_accessor, key)
-    end
+if ENV['TARGET_HOST'].nil? || ENV['TARGET_HOST'] == 'localhost'
+  puts 'Running tests against this machine !'
+  if Gem.win_platform?
+    set :backend, :cmd
+  else
+    set :backend, :exec
   end
+else
+  # load inventory
+  inventory_hash = inventory_hash_from_inventory_file
+  node_config = config_from_node(inventory_hash, ENV['TARGET_HOST'])
 
-  def apply
-    manifest = render.delete("\n")
-    cmd = "bundle exec puppet apply --detailed-exitcodes -e \"#{manifest}\" --modulepath spec/fixtures/modules --libdir lib --debug"
-    result = { output: [], exit_status: nil }
+  if target_in_group(inventory_hash, ENV['TARGET_HOST'], 'docker_nodes')
+    host = ENV['TARGET_HOST']
+    set :backend, :docker
+    set :docker_container, host
+  elsif target_in_group(inventory_hash, ENV['TARGET_HOST'], 'ssh_nodes')
+    set :backend, :ssh
+    options = Net::SSH::Config.for(host)
+    options[:user] = node_config.dig('ssh', 'user') unless node_config.dig('ssh', 'user').nil?
+    options[:port] = node_config.dig('ssh', 'port') unless node_config.dig('ssh', 'port').nil?
+    options[:keys] = node_config.dig('ssh', 'private-key') unless node_config.dig('ssh', 'private-key').nil?
+    options[:password] = node_config.dig('ssh', 'password') unless node_config.dig('ssh', 'password').nil?
+    # Support both net-ssh 4 and 5.
+    # rubocop:disable Metrics/BlockNesting
+    options[:verify_host_key] = if node_config.dig('ssh', 'host-key-check').nil?
+                                  # Fall back to SSH behavior. This variable will only be set in net-ssh 5.3+.
+                                  if @strict_host_key_checking.nil? || @strict_host_key_checking
+                                    Net::SSH::Verifiers::Always.new
+                                  else
+                                    # SSH's behavior with StrictHostKeyChecking=no: adds new keys to known_hosts.
+                                    # If known_hosts points to /dev/null, then equivalent to :never where it
+                                    # accepts any key beacuse they're all new.
+                                    Net::SSH::Verifiers::AcceptNewOrLocalTunnel.new
+                                  end
+                                elsif node_config.dig('ssh', 'host-key-check')
+                                  if defined?(Net::SSH::Verifiers::Always)
+                                    Net::SSH::Verifiers::Always.new
+                                  else
+                                    Net::SSH::Verifiers::Secure.new
+                                  end
+                                elsif defined?(Net::SSH::Verifiers::Never)
+                                  Net::SSH::Verifiers::Never.new
+                                else
+                                  Net::SSH::Verifiers::Null.new
+                                end
+    # rubocop:enable Metrics/BlockNesting
+    host = if ENV['TARGET_HOST'].include?(':')
+             ENV['TARGET_HOST'].split(':').first
+           else
+             ENV['TARGET_HOST']
+           end
+    set :host,        options[:host_name] || host
+    set :ssh_options, options
+    set :request_pty, true
+  elsif target_in_group(inventory_hash, ENV['TARGET_HOST'], 'winrm_nodes')
+    require 'winrm'
 
-    Open3.popen2e(cmd) do |_stdin, stdout_err, wait_thr|
-      while line = stdout_err.gets
-        result[:output].push(line)
-        puts line
-      end
-      result[:exit_status] = wait_thr.value
-    end
+    set :backend, :winrm
+    set :os, family: 'windows'
+    user = node_config.dig('winrm', 'user') unless node_config.dig('winrm', 'user').nil?
+    pass = node_config.dig('winrm', 'password') unless node_config.dig('winrm', 'password').nil?
+    endpoint = "http://#{ENV['TARGET_HOST']}:5985/wsman"
 
-    result
-  end
-
-  def self.to_generalized_data(val)
-    case val
-    when Hash
-      to_generalized_hash_list(val)
-    when Array
-      to_generalized_array_list(val)
-    else
-      val
-    end
-  end
-
-  # returns an array of :k =>, :v => hashes given a Hash
-  # { :a => 'b', :c => 'd' } -> [{:k => 'a', :v => 'b'}, {:k => 'c', :v => 'd'}]
-  def self.to_generalized_hash_list(hash)
-    hash.map { |k, v| { k: k, v: v } }
-  end
-
-  # necessary to build like [{ :values => Array }] rather than [[]] when there
-  # are nested hashes, for the sake of Mustache being able to render
-  # otherwise, simply return the item
-  def self.to_generalized_array_list(arr)
-    arr.map do |item|
-      if item.class == Hash
-        {
-          values: to_generalized_hash_list(item),
-        }
-      else
-        item
-      end
-    end
-  end
-end
-
-class VsphereHelper
-  attr_reader :datacenter
-
-  def initialize
-    credentials = {
-      host: ENV['VCENTER_SERVER'],
-      user: ENV['VCENTER_USER'],
-      password: ENV['VCENTER_PASSWORD'],
-      insecure: true,
+    opts = {
+      user: user,
+      password: pass,
+      endpoint: endpoint,
+      operation_timeout: 300,
     }
-    datacenter = ENV['VCENTER_DATACENTER']
-    @vim = RbVmomi::VIM.connect credentials
-    @datacenter = @vim.serviceInstance.find_datacenter(datacenter)
-  end
 
-  def get_machine(path)
-    local_path = "/#{path.split('/').drop(3).join('/')}"
-    @datacenter.find_vm(local_path)
-  end
-
-  def destroy_machine(path)
-    machine = get_machine(path)
-    if machine
-      machine.PowerOffVM_Task.wait_for_completion if machine.runtime.powerState == 'poweredOn'
-      machine.Destroy_Task.wait_for_completion
-    end
-  end
-
-  def machine_credentials
-    {
-      interactiveSession: false,
-      username: ENV['VCENTER_GUEST_USERNAME'],
-      password: ENV['VCENTER_GUEST_PASSWORD'],
-    }
-  end
-
-  def list_processes(path)
-    machine = get_machine(path)
-    manager = @vim.serviceContent.guestOperationsManager
-    auth = RbVmomi::VIM::NamePasswordAuthentication(machine_credentials)
-    with_retries(max_tries: 10,
-                 max_sleep_seconds: 10,
-                 rescue: NotFinished) do
-      begin
-        manager.authManager.ValidateCredentialsInGuest(vm: machine, auth: auth)
-        manager.processManager.ListProcessesInGuest(vm: machine, auth: auth)
-      rescue RbVmomi::Fault => exception
-        if exception.message.split(':').first == 'GuestOperationsUnavailable'
-          raise NotFinished
-        else
-          raise
-        end
-      end
-    end
-  end
-
-  def execute_command(path, program_path, arguments)
-    machine = get_machine(path)
-    manager = @vim.serviceContent.guestOperationsManager
-    auth = RbVmomi::VIM::NamePasswordAuthentication(machine_credentials)
-    manager.authManager.ValidateCredentialsInGuest(vm: machine, auth: auth)
-    pid = manager.processManager.StartProgramInGuest(vm: machine, auth: auth, spec: { programPath: program_path, arguments: arguments })
-    with_retries(max_tries: 10,
-                 max_sleep_seconds: 10,
-                 rescue: NotFinished) do
-      info = manager.processManager.ListProcessesInGuest(vm: machine, auth: auth, pids: [pid]).first
-      raise NotFinished unless info.exitCode
-      info
-    end
-  end
-end
-
-class TestExecutor
-  def self.shell(cmd)
-    Open3.popen3(cmd) do |_stdin, stdout, stderr, wait_thr|
-      @out = read_stream(stdout)
-      @error = read_stream(stderr)
-      @code = %r{(exit)(\s)(\d+)}.match(wait_thr.value.to_s)[3]
-    end
-    TestExecutor::Response.new(@out, @error, @code, cmd)
-  end
-
-  def self.read_stream(stream)
-    result = ''
-    while line = stream.gets
-      result << line if line.class == String
-      puts line
-    end
-    result
-  end
-
-  # build and apply complex puppet resource commands
-  # the arguement resource is the type of the resource
-  # the opts hash must include a key 'name'
-  def self.puppet_resource(resource, opts = {}, command_flags = '')
-    raise 'A name for the resource must be specified' unless opts[:name]
-    cmd = "puppet resource #{resource} "
-    options = ''
-    opts.each do |k, v|
-      if k.to_s == 'name'
-        @name = v
-      else
-        options << "#{k}=#{v} "
-      end
-    end
-    cmd << "#{@name} "
-    cmd << options
-    cmd << " --libdir lib --modulepath spec/fixtures/modules #{command_flags}"
-    # apply the command
-    response = shell(cmd)
-    response
-  end
-end
-
-class TestExecutor::Response
-  attr_reader :stdout, :stderr, :exit_code, :command
-
-  def initialize(standard_out, standard_error, exit, cmd)
-    @stdout = standard_out
-    @stderr = standard_error
-    @exit_code = exit
-    @command = cmd
+    winrm = WinRM::Connection.new opts
+    Specinfra.configuration.winrm = winrm
   end
 end
